+++ QA_Challenge/gallery.md
@@ -1,0 +1,23 @@
+![alt text](images/Limehome_QA_Challenge_Test_Report_Solution.png)
+
+#### Test report as at final edits Friday Aug 8th. (I ran some more test just to double check)
+
+---
+
+<br><br><br>
+
+![alt text](images/Limehome_QA_Challenge_Test_Report_Solution_playwright_View.png)
+
+#### Similar test report as as generated by playwright
+
+---
+
+<br><br><br>
+
+![alt text](images/Limehome_QA_Challenge_Test_Report_Solution_Excel_Summary.png)
+
+#### Test Scenario overview. Related file can be found in the parent folder
+
+---
+
+<br><br><br>
