+++ data/i18n/index.ts
@@ -1,6 +1,10 @@
 import { en } from './en';
 import { de } from './de';
 
+/**
+ * Application language strings
+ * Avoid using language strings as test selectors, allowed in very few cases.
+ */
 export const language = {
   en,
   de
