+++ readme.md
@@ -1,23 +1,28 @@
 # Task Summary.
+![alt text](QA_Challenge/images/Limehome_QA_Challenge_Test_Report_Solution.png)
 
-- Everything below this section are technical outlines. This section focuses on task summary and other relevant documents.
-- For Test strategy document, please see `` PDF under /QA_Challenge
-  - This document also contains some bugs/issues I found on the Frontend
-- For test scenario breakdown, please see `` sheet under /QA_Challenge
+Everything below this section are technical outlines. This section focuses on task summary and other relevant documents.
+- For Test strategy document, please see `Limehome challenge - QA Engineer. Approach strategy. Chris Enitan` PDF under /QA_Challenge
+  - This document also contains some bugs or issues I found while working on the Frontend
+- For test scenario breakdown, please see `Limehome challenge - QA Engineer. Test Scenarios. Chris Enitan` sheet under /QA_Challenge
+  - I shared an efficiency concern & approach about this sheet on [feedbacks](./QA_Challenge/feedbacks.md) point 3.
 - For details on the conventions I used, please see [contributions](./contributing.md)
 - There is a Feedback document tracking comments on my progress at hourly intervals. See [feedbacks](./QA_Challenge/feedbacks.md)
 - Result for my latest test run can be viewed before overwriting with another test run. `npm run report`
+  - I added a visual summary for this on [Gallery](./QA_Challenge/gallery.md)
+- The project is large because I included the playwright report data from my last run.
+
 
 # Introduction
 
-API & End to End tests for the [Limehome UI](https://www.limehome.com/) [Limehome API](https://api.limehome.com/properties/v1/public/).
+API & End to End tests challenge solutions for the [Limehome UI](https://www.limehome.com/) [Limehome API](https://api.limehome.com/properties/v1/public/).
 
 ## Project preconditions
 
 - Target Environments: https://www.limehome.com/en and https://api.limehome.com/properties/v1/public/
 - Tests are by default executed against the Applications English language variant except otherwise stated.
 - Tests are organized in application page hierarchy. See [Project Structure](#project-structure) for details
-- Test default timeout is `20000` milliseconds. See the playwright config file for other boundaries.
+- Test default test timeout is `20000` milliseconds and assertion timeout is set to `5000` ms. See the playwright config file for other boundaries.
 - Device folder does not constrain test structure, we may consider re-running some desktop test in a different viewport as supported by Playwright.
 
 #### System Requirements 
@@ -26,6 +31,7 @@ API & End to End tests for the [Limehome UI](https://www.limehome.com/) [Limehom
 ---
 
 ## ENV Variables 
+You don't need any of these, they are defaulted. But I added the logics to demonstrate easy future adaption.
 Set these with the usual Node process env
 |Space|Possible Values|
 |--|--|
@@ -42,6 +48,7 @@ Limehome-test-management
 		└───Playwright Fixture		- Folder organization according to Application Pages
 		└───Pages Models			- Folder organization according to Application Pages
 ├───data					- i18n, CSS selectors, files and other static assets
+	└───api_schemas				- API Validation Schemas
 └───test                    - Main tests directory
   	└───Project  				- API/UI-E2E etc
 	  	└───Device  				- Mobile/Desktop/Pad etc (Currently only contains Desktop)
@@ -68,7 +75,7 @@ There are currently no existing accounts needed on the solution but all static t
    - Lint and format the project:  `npm run check`
 
 
-## Running Test Steps
+## Steps for running Test
 
 To run tests, use the default playwright commands. Tests are organized in scenario and configured in playwright projects.
 
@@ -77,10 +84,10 @@ To run tests, use the default playwright commands. Tests are organized in scenar
 - Run via the Playwright extension [ms-playwright.playwright](https://marketplace.visualstudio.com/items?itemName=ms-playwright.playwright)
 
 ## Test Reports
-
 - HTML Test reports are automatically generated by [Playwright Reporter](https://playwright.dev/docs/trace-viewer-intro#opening-the-html-report)
   - You can run `npm run report` to view them
   - The HTML can also be found under `/playwright/index.html`
+- Test run via the terminal also print out reports in standard list format
 
 
 ## Debugging
