+++ services/api/custom_api_assertions.ts
@@ -1,5 +1,5 @@
 import Ajv from 'ajv';
-const ajv = new Ajv(); // options can be passed, e.g. {allErrors: true}
+const ajv = new Ajv();
 
 export const ApiAssertions = {
   /**
