+++ types/index.ts
@@ -7,10 +7,10 @@ type AllowedDigits = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9';
 // Using this limited scope for now
 type Year = `${19 | 20}${AllowedDigits}${AllowedDigits}`;
 
-// Month: 01 to 12
+// Normal months no indexing
 type ValidMonths = '01' | '02' | '03' | '04' | '05' | '06' | '07' | '08' | '09' | '10' | '11' | '12';
 
-// Day: 01 to 31
+// Normal days no indexing
 type Days =
   | '01'
   | '02'
